+++ ProfileMetaData.cpp
@@ -28,7 +28,7 @@ using namespace FiftyoneDegrees::Common;
 ProfileMetaData::ProfileMetaData()
     : ProfileMetaData(
         0, // profile id
-        -1 // component id
+        (byte)-1 // component id
     ) { }
 
 ProfileMetaData::ProfileMetaData(
